+++ kotgh/src/api/auth.ts
@@ -1,0 +1,22 @@
+export const fetchData = async () => {
+    const response = await fetch('/api/data'); // Fetch data from the API
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+    return response.json(); // Parse and return JSON
+  };
+  
+  export const postData = async (newData: string) => {
+    const response = await fetch('/api/data', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json', // Specify JSON content
+      },
+      body: JSON.stringify({ newData }),
+    });
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+    return response.json();
+  };
+
