+++ server/README.md
@@ -1,1 +1,3 @@
 `npm start`
+
+`npm run dev` - to update with changes from .ts file
