+++ server/package.json
@@ -6,7 +6,7 @@
   "main": "dist/server.js",
   "scripts": {
     "start": "npm run build && node dist/server.js",
-    "dev": "ts-node-dev --respawn --transpile-only src/server.ts",
+    "dev": "node --experimental-specifier-resolution=node --loader ts-node/esm ./src/server.ts",
     "build": "tsc"
   },
   "keywords": [],
@@ -18,7 +18,9 @@
     "express": "^4.19.2",
     "helmet": "^7.1.0",
     "node-fetch": "^3.3.2",
+    "nodemon": "^3.1.9",
     "path": "^0.12.7",
+    "ts-node": "^10.9.2",
     "url": "^0.11.3"
   },
   "devDependencies": {
